% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_quarto.R
\name{build_quarto}
\alias{build_quarto}
\title{Provided a session number, generates a quarto header and then one chunk for each milestone sharing the same session number}
\usage{
build_quarto(session_n)
}
\arguments{
\item{session_n}{}
}
\value{
quarto headers and chunks
}
\description{
Provided a session number, generates a quarto header and then one chunk for each milestone sharing the same session number
}
