% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bullet_list.R
\name{bullet_list}
\alias{bullet_list}
\title{Convert a vector into a bulletted quarto list with less agony}
\usage{
bullet_list(...)
}
\arguments{
\item{...}{vector}
}
\value{
lump of quarto
}
\description{
Convert a vector into a bulletted quarto list with less agony
}
\examples{
bullet_list(1:4)
}
