% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chilli_adder.R
\name{chilli_adder}
\alias{chilli_adder}
\title{chilli_adder}
\usage{
chilli_adder(str, html = T)
}
\arguments{
\item{str}{\itemize{
\item an input string describing the training level
}}

\item{html}{\itemize{
\item HTML-friendly output required
}}
}
\value{
\itemize{
\item standardised training-level string with icon
}
}
\description{
chilli_adder
}
\examples{
chilli_adder("beginner")
chilli_adder("intermediate", html = FALSE)
}
