% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_lines.R
\name{clean_lines}
\alias{clean_lines}
\title{Cleans out empty lines at head and foot of files}
\usage{
clean_lines(file)
}
\arguments{
\item{file}{}
}
\description{
Cleans out empty lines at head and foot of files
}
