% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback_url.R
\name{feedback_url}
\alias{feedback_url}
\title{Get a feedback URL from a session title and date}
\usage{
feedback_url(session, date, type = "training")
}
\arguments{
\item{session}{A session title}

\item{date}{A string in YYYY-MM-DD format}

\item{type}{Type of training}
}
\value{
a feedback url
}
\description{
Get a feedback URL from a session title and date
}
\examples{
feedback_url("KIND one-off training session: why bother with…Power BI?", "2024-06-13")
}
