% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_finder.R
\name{file_finder}
\alias{file_finder}
\title{Find file based on session number, path, and pattern}
\usage{
file_finder(session_n, path, pattern)
}
\arguments{
\item{pattern}{}
}
\description{
Find file based on session number, path, and pattern
}
