% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_milestones.R
\name{join_milestones}
\alias{join_milestones}
\title{Joins the milestones to their appropriate chunk in the quarto}
\usage{
join_milestones(session_n)
}
\arguments{
\item{session_n}{}
}
\value{
quarto file
}
\description{
Joins the milestones to their appropriate chunk in the quarto
}
\examples{
join_milestones(1)
}
