% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kind_links.R
\name{kind_links}
\alias{kind_links}
\title{KIND links}
\usage{
kind_links()
}
\value{
an Rmarkdown/Quarto text slug containing KIND network info and links
}
\description{
KIND links
}
\examples{
kind_links()
}
