% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chunk.R
\name{make_chunk}
\alias{make_chunk}
\title{Provided a milestone number, generates a quarto chunk for that milestone}
\usage{
make_chunk(milestone_number)
}
\arguments{
\item{milestone_number}{}
}
\description{
Provided a milestone number, generates a quarto chunk for that milestone
}
