% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qmd.R
\name{make_qmd}
\alias{make_qmd}
\title{Given a training session number, builds and opens a .qmd from a text section file and a milestones file}
\usage{
make_qmd(session_n, render = FALSE)
}
\arguments{
\item{render}{}
}
\value{
qmd and html version of training file
}
\description{
Given a training session number, builds and opens a .qmd from a text section file and a milestones file
}
