% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milestone_split.R
\name{milestone_split}
\alias{milestone_split}
\title{Given a filename, this function splits the script into sequentially-numbered milestone files. The aim is to allow all the code for a session to be written in one script, and then split using this function into free-standing scripts, allowing trainees to see what the end-game is at each stage of the training}
\usage{
milestone_split(session_n)
}
\arguments{
\item{session_n}{}
}
\description{
Given a filename, this function splits the script into sequentially-numbered milestone files. The aim is to allow all the code for a session to be written in one script, and then split using this function into free-standing scripts, allowing trainees to see what the end-game is at each stage of the training
}
