% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_date.R
\name{nice_date}
\alias{nice_date}
\title{Standard date format for KIND work}
\usage{
nice_date(date)
}
\arguments{
\item{date}{A date- or Posix class date/date time}
}
\value{
A text string in the format (10:00) Mon 10th June 2024
}
\description{
Standard date format for KIND work
}
\examples{
nice_date(today())
}
