% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_from_email.R
\name{org_from_email}
\alias{org_from_email}
\title{Generate an organisation from supplied email addresses}
\usage{
org_from_email(df, from = "Email", to = "org_from_email", lump_misc = TRUE)
}
\arguments{
\item{df}{A tibble containing email addresses}

\item{from}{Column containing email addresses}

\item{to}{Column to create containing infered organisations}

\item{lump_misc}{Option to gather non-standard addresses into an 'Other' category, or to retain their stem (e.g. nhs.scot).}
}
\value{
A tibble
}
\description{
Generate an organisation from supplied email addresses
}
\examples{
data \%>\%
org_from_email()
}
