% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_from_free.R
\name{org_from_free}
\alias{org_from_free}
\title{Infers organisation name from free text}
\usage{
org_from_free(df, from = "Organisation", to = "org_from_free")
}
\arguments{
\item{df}{Tibble containing free text}

\item{from}{Column containing free text}

\item{to}{Column to create containing organisation name}
}
\value{
A tibble
}
\description{
Infers organisation name from free text
}
\examples{
data \%>\% org_from_free()
}
