% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_from_session.R
\name{paths_from_session}
\alias{paths_from_session}
\title{Take a session # and return paths to relevant files}
\usage{
paths_from_session(session_n)
}
\arguments{
\item{session_n}{}
}
\description{
Take a session # and return paths to relevant files
}
