% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation_proj.R
\name{presentation_proj}
\alias{presentation_proj}
\title{Create standardised KIND slides}
\usage{
presentation_proj(title)
}
\value{
A project folder containing boilerplate training materials
}
\description{
Creates standard training slides as a RevealJS Quarto presentation. Intended for VScode, rather than Rstudio. To be run inside intended containing folder.
}
\examples{
presentation_proj("Neural Nets made ridiculously simple")
}
