% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_teams_att.R
\name{read_teams_att}
\alias{read_teams_att}
\title{Easily read Microsoft Teams attendance files}
\usage{
read_teams_att(fn, organiser_name = "Brendan Clarke")
}
\arguments{
\item{fn}{A filename of a Teams attendance .csv, or path of a folder containing several .csvs}

\item{organiser_name}{The name of the organiser, which can optionally be removed from the data}
}
\value{
A tibble
}
\description{
\code{read_teams_att()} reads Microsoft Teams attendance files, guesses the encoding, and produces a tidy tibble of attendees. Teams attendance files are painful to read in R. First, they use UTF16LE encoding, which causes problems with standard methods. Next, the structure of the .csv file itself is repetitious, requiring some careful tidying. Finally, they are tab-demarcated, but saved as .csv, causing difficulties for other methods that infer separators.
}
\examples{
df <- read_teams_att("data/jan_2023_meeting.csv")
}
