% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotcodes.R
\name{scotcodes}
\alias{scotcodes}
\title{scotcodes}
\usage{
scotcodes(from_var, from_val, to_var)
}
\arguments{
\item{from_var}{\itemize{
\item an unquoted column name of the variable that you want to convert from
}}

\item{from_val}{\itemize{
\item a quoted string containing the value to look up in that column
}}

\item{to_var}{\itemize{
\item a quoted string containing full of partial
}}
}
\value{
tibble containing matching column(s)
}
\description{
Geography codes are important but awkward. This function takes any of a range of possible geography codes, and returns whatever corresponding geography codes you need. The data contains both _Code and _Name variants of the following geography codes:
\itemize{
\item DZ2011 - datazones
\item IZ2011 - intermediate zones
\item MMWard - wards
\item LA - local authorities
\item SPC
\item UKPC
\item UKPC - parliamentary consistuencies
\item HB - health boards
\item HIA - health integration areas
\item SPD
}
}
\examples{
scotcodes(HB_Name, "Highland", "DZ")
scotcodes(HB_Name, c("Highland", "Grampian"), "DZ2011_Name")
scotcodes(HB_Code, "S08000020", "Name") # returns all available data for Grampian

}
