% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotmap.R
\name{scotmap}
\alias{scotmap}
\title{scotmap
easily generate sf data for Scotland, based on health boards, health integration areas, intermediate zones, or data zones}
\usage{
scotmap(type = "DZ", filter = NA, value = "All")
}
\arguments{
\item{type}{The type of sf data needed}

\item{subgroup}{How to filter the sf data}
}
\value{
sf object
}
\description{
scotmap
easily generate sf data for Scotland, based on health boards, health integration areas, intermediate zones, or data zones
}
\examples{
scotmap() # gives national data at the DZ2011 level
scotmap("IZ", "name", "hill") # interzones with "hill" in the name
}
