% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_from_email.R
\name{sector_from_email}
\alias{sector_from_email}
\title{Infer sector from supplied email address}
\usage{
sector_from_email(df, from = "Email", to = "sector_from_email")
}
\arguments{
\item{df}{A tibble containing email addresses}

\item{from}{A column name containing email addresses}

\item{to}{A column name as the destination for sector descriptions}
}
\value{
A tibble
}
\description{
Infer sector from supplied email address
}
\examples{
data \%>\% sector_from_email()

}
