% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surviving_daughters_stationary.R
\name{surviving_daughters_stationary}
\alias{surviving_daughters_stationary}
\title{Surviving daughters (stationary popualtion)}
\usage{
surviving_daughters_stationary(df, age_a, ffab = 0.4886)
}
\arguments{
\item{df}{dataframe containing Lx and Fx values by age}

\item{age_a}{age of mother}

\item{ffab}{fraction female at birth, default is 0.4886.}
}
\value{
a number representing the implied number of surviving daughters
}
\description{
Given Lx and Fx, calculate the implied number of surviving daughters assuming a stationary population
}
