% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams_transcript_offset.R
\name{teams_transcript_offset}
\alias{teams_transcript_offset}
\title{Apply an offset to fix the timings of subtitles in Teams recordings
#'}
\usage{
teams_transcript_offset(input, offset = 9)
}
\arguments{
\item{input}{.vtt transcript from Teams}

\item{offset}{an offset in seconds. Using 9 here makes subtitles appear 9 seconds earlier.}
}
\value{
.vtt transcript
}
\description{
Apply an offset to fix the timings of subtitles in Teams recordings
#'
}
\examples{

source_path = "C:\\\\Users\\\\BrendanCl\\\\2. Open Source in NHS Lothian.vtt"

readr::read_lines(source_path) |>
 stringr::str_replace_all("\\\\d\\\\d:\\\\d\\\\d:\\\\d\\\\d.\\\\d\\\\d\\\\d", \(x) teams_transcript_offset(input = x, offset = 2)) |>
 readr::write_lines(paste0(tools::file_path_sans_ext(source_path), "_fixed.vtt"))
}
