% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_sessions.R
\name{training_sessions}
\alias{training_sessions}
\title{Find current training for a particular platform}
\usage{
training_sessions(
  tr_type = "all",
  start_date = "today",
  end_date = lubridate::dmy("01-12-2025")
)
}
\arguments{
\item{tr_type}{Training type}

\item{start_date}{A ymd-formatted date}

\item{end_date}{A ymd-formatted date}
}
\value{
Tibble with Session (with link), Date, Level
}
\description{
Find current training for a particular platform
}
\examples{
training_sessions("R")
training_sessions("Excel", "01-07-2024")

}
