% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_finder.R
\name{file_finder}
\alias{file_finder}
\title{Find file based on session number, path, and pattern}
\usage{
file_finder(session_n, path, pattern)
}
\arguments{
\item{session_n}{numeric session number}

\item{path}{root file path}

\item{pattern}{3-digit file extension}
}
\value{
file path
}
\description{
Find file based on session number, path, and pattern
}
\examples{
file_finder(1, "/example", ".Rmd")
}
