% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_date.R
\name{nice_date}
\alias{nice_date}
\title{Make a friendly date from date/datetime}
\usage{
nice_date(date)
}
\arguments{
\item{date}{A date/datetime}
}
\value{
a nice (and standard for KIND) date string in the format "Tue 9th September 2025"
}
\description{
Make a friendly date from date/datetime
}
\examples{
nice_date(lubridate::today())
}
