% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_sessions.R
\name{training_sessions}
\alias{training_sessions}
\title{Find current training for a particular platform}
\usage{
training_sessions(
  tr_type = "all",
  start_date = "today",
  end_date = "2026-12-31",
  session_level = "all",
  hide_area = FALSE,
  n = 0,
  output_type = "kableExtra"
)
}
\arguments{
\item{tr_type}{Training type}

\item{start_date}{A ymd-formatted date. We add 1 day to the date, so that you don't have to.}

\item{end_date}{A ymd-formatted date}

\item{session_level}{Training level required}

\item{hide_area}{Hide the area column?}

\item{n}{Number of sessions required}

\item{output_type}{Output type required: pick from kableExtra, tibble, DT}
}
\value{
Tibble with Session (with link), Date, Level
}
\description{
Find current training for a particular platform
}
\examples{
training_sessions("R")
training_sessions("Excel", "2024-07-01")

}
