% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimeo_thumb.R
\name{vimeo_thumb}
\alias{vimeo_thumb}
\title{Make a thumbnail from title and speaker information for Vimeo}
\usage{
vimeo_thumb(
  topic,
  speaker,
  logo =
    "https://raw.githubusercontent.com/NES-DEW/dew-img/refs/heads/main/kind/kind_banner_700.png"
)
}
\arguments{
\item{topic}{The session topic}

\item{speaker}{Speaker name (and affiliation)}

\item{logo}{A direct URL for any custom logo, e.g. for theme weeks}
}
\value{
saves png thumbnail in working directory
}
\description{
Make a thumbnail from title and speaker information for Vimeo
}
\examples{

vimeo_thumb("Open Source in NHS Lothian", "Zena Trendell (NHS Lothian)  \nJules Augley (NHS Lothian)", "https://raw.githubusercontent.com/NES-DEW/dew-img/refs/heads/main/kind/kind_banner_700_osw.png")
}
